requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0

# API & Frontend
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
python-multipart>=0.0.6

# Experiment Tracking
mlflow>=2.9.0

# Data & Visualization
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
Pillow>=10.0.0
plotly>=5.18.0

# Config & Utilities
pyyaml>=6.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0

# Linting
ruff>=0.1.0
